requirements.txt
aiobotocore
aiohttp
marshmallow
marshmallow-dataclass
python-decouple
aiosqlite

